01_LLM_API/requirements.txt
accelerate==1.8.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
airportsdata==20250706
annotated-types==0.7.0
anyio==4.9.0
artifacts-keyring==0.4.0
astor==0.8.1
attrs==25.3.0
backports.tarfile==1.2.0
blake3==1.0.5
boto3==1.38.33
botocore==1.38.33
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.10.2
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.3
cupy-cuda12x==13.5.1
dataclasses-json==0.6.7
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
einops==0.8.1
email_validator==2.2.0
etelemetry==0.3.1
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastrlock==0.8.3
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
frozenlist==1.7.0
fsspec==2025.5.1
gguf==0.17.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-aiplatform==1.101.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.24.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
hyperframe==6.1.0
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
langchain==0.3.26
langchain-aws==0.2.23
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-genai==2.1.4
langchain-google-vertexai==2.0.27
langchain-naver==0.1.0rc3
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langsmith==0.3.45
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
milvus-lite==2.4.12
mistral_common==1.8.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.2.21
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numba==0.61.2
numexpr==2.11.0
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.85.0
opencv-python-headless==4.12.0.88
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.3.0
partial-json-parser==0.2.1.1.post6
pathlib==1.0.1
pillow==11.2.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil==7.0.0
puremagic==1.29
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
pymilvus==2.5.11
PyMuPDF==1.26.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.0
ray==2.48.0
rdflib==7.1.4
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scipy==1.14.1
SecretStorage==3.3.3
sentencepiece==0.2.0
sentry-sdk==2.33.0
setuptools==79.0.1
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.21.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traits==7.0.2
transformers==4.52.4
triton==3.3.0
twine==6.1.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
vllm==0.9.2
watchfiles==1.1.0
websockets==15.0.1
xformers==0.0.30
xgrammar==0.1.19
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0